Login screen UI

+++ Twitteroid/TWRTwitterAPIManager+TWRFeed.m
@@ -13,7 +13,7 @@
 
 - (void)getFeedOlderThatTwitID:(NSString *)twitID count:(NSUInteger)count completion:(void(^)(NSError *error, NSArray *items))completion
 {
-    [self.twitter getStatusesHomeTimelineWithCount:[NSString stringWithFormat:@"%d", count] sinceID:nil maxID:twitID trimUser:nil excludeReplies:nil contributorDetails:nil includeEntities:nil successBlock:^(NSArray *statuses) {
+    [self.twitter getStatusesHomeTimelineWithCount:[NSString stringWithFormat:@"%lu", (unsigned long)count] sinceID:nil maxID:twitID trimUser:nil excludeReplies:nil contributorDetails:nil includeEntities:nil successBlock:^(NSArray *statuses) {
         completion(nil, statuses);
     } errorBlock:^(NSError *error) {
         completion(error, nil);
